+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- fix(s3): fix content type if not guessed [#18](https://github.com/etalab/flask-storage/pull/18)
 
 ## 1.4.4 (2026-03-12)